Rename package chunk to chunker
