+++ chunker/chunker.go
@@ -1,0 +1,5 @@
+package chunker
+
+// Chunker ...
+type Chunker struct {
+}
