+++ chunker/header.go
@@ -1,4 +1,4 @@
-package chunk
+package chunker
 
 import (
 	"bytes"
