+++ chunker/header_test.go
@@ -1,4 +1,4 @@
-package chunk
+package chunker
 
 import (
 	"bytes"
